refactor: app 파일 위치 및 컴포넌트 수정
