+++ src/app/main.tsx
@@ -1,0 +1,11 @@
+import { StrictMode } from "react";
+
+import { createRoot } from "react-dom/client";
+
+import App from "./App.tsx";
+
+createRoot(document.getElementById("root")!).render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+);
