--- src/index.css
@@ -1,4 +1,0 @@
-html, body {
-  background: #fff;
-  color: #000;
-}
